integrate teleoperation client with ros node, show odometry data on the web app
(cherry picked from commit b6b996b19b43fbf562c12306771f38448f95c229)

+++ pc_client/teleop_actions/StreamVideoAction.cpp
@@ -24,7 +24,7 @@ void StreamVideoAction::init() {
 
 bool StreamVideoAction::loop() {
     if ((len = recv(sock, buffer, BUFFER_SIZE, 0)) >= 0) {
-        if (len < sizeof(rtc::RtpHeader))
+        if (static_cast<long unsigned int>(len) < sizeof(rtc::RtpHeader))
             return true;
 
         auto rtp = reinterpret_cast<rtc::RtpHeader *>(buffer);
